User entitiy (#69)
* feat: 유저 엔티티 수정 및 유저 요청 api 수정

stat, status, friends, blocks 추가

* refactor: enum status 를 string 형식으로 수정

+++ backend/app/src/auth/two-factor.service.ts
@@ -35,7 +35,7 @@ export class TwoFactorService {
   async disable(uid: number): Promise<void> {
     const user = await this.userRepository.findOneBy({ uid })
     user.twoFactor = false
-    await this.twoFactorRepository.delete({ user })
+    await this.twoFactorRepository.delete({ user: { uid } })
     await this.userRepository.save(user)
   }
